set cookies and created cart in base and cart.js

+++ shop/models.py
@@ -20,7 +20,7 @@ class Product(models.Model):
     """ when a customer order a product """
 
     name = models.CharField(max_length=200)
-    price = models.FloatField()
+    price = models.DecimalField(max_digits=7, decimal_places=2)
     digital = models.BooleanField(default=False, null=True, blank=True)
     image = models.ImageField(null=True, blank=True)
